+++ daemon/src/services/server.rs
@@ -395,6 +395,7 @@ impl DaemonControl for DaemonService {
 
         debug!("Successfully reloaded snippets and settings into service.");
         taurine_core::settings::set_cached_audio_theme(settings.audio_theme);
+        taurine_core::settings::set_cached_audio_volume(settings.audio_volume);
         tokio::spawn(crate::dictionary_manager::check_and_update_dictionary());
         Ok(Response::new(ReloadResponse { success: true }))
     }